add command processor
 - ReaperCommandProcessor waits for commands and executes them. Initial
   command set is "list" to list CD card and "exit" which prints exiting.
   Unrecognized commands are ignored.

 - CommandLoop is example sketch that uses CommandProcessor

 - python sdreaper can run one or more commands, for example:

   sdreaper -p /dev/tty.usbmodem1421 -b 9600 list list exit

   runs "list" twice then "exit" and prints results

  - python currently prints some protocol debugging and the Arduino
    also prints debugging to the debug (Serial) port. Need to make
    this configurable in python and a compile time option in Reaper.

Command protocol summary:

  AR = Arduino Reaper
  PC = Python Client

A command line is an ASCI sequence of up to 127 bytes followed by a LF (0x0A)
A command line has the form:

  verb<SP>argument<LF>

  verb - identifies action (e.g., list)
  argument - optional, typically a file name to operate on

On receiving a command AR ignores input until processing of the command is
complete. Commands may produce output. The end of a command response
is always signaled by an EOT (0x04).

While listening for commands if AR receives a SYN (0x16) it responds with a
SYN. This allows the PC to know that AR is listening.

I experienced a lot of dropped bytes when reading from SerialUSB on the
Arduino. A little googling indicates others have run into this too. Some
causes:

 - not using available() properly
 - mismatch between serial settings (baud, parity, stop bits)
 - baud skew or drift between Arduino and laptop

I think I was seeing some skew as I was losing characters with the IDE Serial
Monitor as well as example test scripts like this one:

  http://www.gammon.com.au/serial

Adding some python delays and the SYN handshake seems to have cleared things
up. I'll write a test script to exercise this more fully.

TODOs:
  - doxygen for Reaper libraray including command protocol
  - test sketch and python script for protocol reliability
  - file transfer