smarter python speed test
It waits for USB connection so can be started before Arduino.

It reads 16K blocks with no timeout and computes timing after Ctrl-C is
received. Run it and press Ctrl-C after it stops printing '.' chars. It is ok
to wait longer as only actual read time is measured.

With an Arduino Zero and Arduino.cc toolchain, numbers are much better:

% python serial-read.py /dev/tty.usbmodem1421 data.out
.connected

reading data, printing a dot "." every 10240 bytes
....................^C
read 102400 bytes in 0.1510605239891447 seconds
rate = 677873.9891525764

% od -c data.out
0000000    1   0   2   4   0   0  \r  \n   x   x   x   x   x   x   x   x
0000020    x   x   x   x   x   x   x   x   x   x   x   x   x   x   x   x
*
0310000

% head -1 data.out
102400