fix #1 - handle aborted transfers
Add a software watchdog timer for file transfers. During send()
if the receiver has not ACKed or NAKed in the last 3 seconds, cancel the send().
Typical transfer rates are 100 packets per second and the receiver ACKs in
response to each packet, so 3 seconds of silence between two packets is a
good indicator that the receiver disconnected.