close #3 - blink led during file transfer
Three 100ms flashes at start of transfer and then toggle after every 20
blocks sent.

Note: SAMD boards do not appear to support reading a pin that is set to OUTPUT
mode. This is unlike AVR boards such as the UNO where it is a standard
practice. So we keep track of the state when blinking.