add simple device name table
Database is editable text file named devices.txt in top level of data
directory. Each line is:

name = device_id

where name is the friendly name to use for the device and deivce_id is the
SAMD chip id.

The database is created if needed at startup, and new devices are added
with an abbreviation of the chip id as their name.

Next: allow editing of name in app