auto-detect device port
If no port is specified sdreaper will look for an Arduino Reaper device
connected to a USB port and use the first device found. If no device is found
an explanatory message is printed.

If a port is specified the port will be used. In this case sdreaper will
wait indefinitely until the port is available.